save will now update the 'articles' database, and render articles to the DOM
