+++ src/scripts/articles/articleHTMLManager.js
@@ -1,6 +1,5 @@
-const articleHTMLManager = {
-  newArticleForm: () => {
-    return `
+const newArticleForm = () => {
+  return `
     <form action="">
     <legend><b>News Article</b></legend>
         <input type="hidden" id="entryId" value="" />
@@ -21,27 +20,6 @@ const articleHTMLManager = {
         </fieldset>
         
     `;
-  },
-  makeNewsArticleComponent: newsArticle => {
-    return `
-      <h1>${newsArticle.title}</h1>
-      <section>${newsArticle.synopsis}</section>
-      <article>${newsArticle.url}</article>
-      <article>${newsArticle.timestamp}</article>
-      <button id="editNewsArticle--${newsArticle.id}" class="editBtn">Edit</button>
-      <button id="deleteNewsArticle--${newsArticle.id}" class="deleteBtn">Delete</button>
-      `;
-  },
-  renderArticles: articles => {
-    const articlesDom = document.getElementById("newsArticles");
-  
-    articlesDom.innerHTML = "";
-  
-    articles.forEach(article => {
-    articlesDom.innerHTML += makeNewsArticleComponent(article)    
-    });
-  }
 };
 
-
-export default articleHTMLManager;
+export default newArticleForm;
